% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main-function.R
\name{coastal_wave}
\alias{coastal_wave}
\title{Calculates coastal wave metrics}
\usage{
coastal_wave(type = "fetch", ...)
}
\arguments{
\item{type}{one of "fetch", "direction", or "orientation"}

\item{...}{additional parameters for the wave metrics functions}

\item{land1}{IntegerMatrix representing the primary land area.}

\item{land2}{IntegerMatrix representing the secondary land area used for alternative mapping.}

\item{dx}{Integer specifying the search distance for near-coast detection.}

\item{dwx}{Integer controlling the magnitude of the random search distance.}

\item{cl}{Integer indicating the coastal cell class to process.}

\item{nsearches}{Integer specifying the number of iterations for the random search.}

\item{land1llx}{Double: lower left x-coordinate for land1.}

\item{land1uly}{Double: upper left y-coordinate for land1.}

\item{land1cellszx}{Double: cell size in the x-direction for land1.}

\item{land1cellszy}{Double: cell size in the y-direction for land1.}

\item{land2llx}{Double: lower left x-coordinate for land2.}

\item{land2uly}{Double: upper left y-coordinate for land2.}

\item{land2cellszx}{Double: cell size in the x-direction for land2.}

\item{land2cellszy}{Double: cell size in the y-direction for land2.}

\item{verbose}{Integer print messages to track progress. Use 0 for no messages, 1 for progress bar or 2 for text messages.}
}
\value{
A matrix with the computed wave fetch metric based on the log-transformed search approach.
}
\description{
Calculates coastal wave metrics
}
\details{
This function is a wrapper around the three individual functions used to calculate
the metrics.
}
\examples{
\dontrun{
cw_result <- coastal_wave(type = "fetch", 
 land1 = land1_matrix,
 land2 = land2_matrix, 
 dx = 50, dwx = 2000, cl = 2, jit = 1,
 land1llx, land1uly, land1cellszx, land1cellszy,
 land2llx, land2uly, land2cellszx, land2cellszy)
}

}
