% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coastal_wave_orientation}
\alias{coastal_wave_orientation}
\title{Compute the wave orientation metric using log-transformed distances.}
\usage{
coastal_wave_orientation(
  land1,
  land2,
  dx,
  dwx,
  cl,
  jit,
  land1llx,
  land1uly,
  land1cellszx,
  land1cellszy,
  land2llx,
  land2uly,
  land2cellszx,
  land2cellszy,
  verbose = 1L
)
}
\arguments{
\item{land1}{IntegerMatrix representing the primary land area.}

\item{land2}{IntegerMatrix representing the secondary land area used for alternative mapping.}

\item{dx}{Integer specifying the search distance for near-coast detection.}

\item{dwx}{Integer controlling the magnitude of the random search distance.}

\item{cl}{Integer indicating the coastal cell class to process.}

\item{land1llx}{Double: lower left x-coordinate for land1.}

\item{land1uly}{Double: upper left y-coordinate for land1.}

\item{land1cellszx}{Double: cell size in the x-direction for land1.}

\item{land1cellszy}{Double: cell size in the y-direction for land1.}

\item{land2llx}{Double: lower left x-coordinate for land2.}

\item{land2uly}{Double: upper left y-coordinate for land2.}

\item{land2cellszx}{Double: cell size in the x-direction for land2.}

\item{land2cellszy}{Double: cell size in the y-direction for land2.}

\item{verbose}{Integer print messages to track progress. Use 0 for no messages, 1 for progress bar or 2 for text messages.}

\item{nsearches}{Integer specifying the number of iterations for the random search.}
}
\value{
A NumericMatrix with the computed wave fetch metric based on the log-transformed search approach.
}
\description{
This function calculates a wave orientation metric (wx32) for coastal cells,
and uses a log-transformed approach for the random distance search.
An alternative mapping search is performed when the primary search is out-of-bound.
}
\details{
The original name of this function was \code{coastwx32twomaplatlono}
}
