% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isitnearcoast}
\alias{isitnearcoast}
\title{Identify near-coast cells based on a land matrix and a search distance.}
\usage{
isitnearcoast(land, dx)
}
\arguments{
\item{land}{IntegerMatrix representing land (1) and water (0).}

\item{dx}{Integer specifying the search distance (in cells).}
}
\value{
An IntegerMatrix with coastal cells marked as 1 and near-coast cells marked as 2.
}
\description{
For cells identified as coastal, this function marks surrounding water cells (value 0)
within a distance dx as near-coast (value 2).
}
