% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land1}
\alias{land1}
\title{Land raster at high resolution}
\format{
\subsection{\code{land1}}{

A serialized raster.
To unpack use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(terra)
data(land1)
land1 <- unwrap(land1)
}\if{html}{\out{</div>}}
}
}
\source{
authors
}
\usage{
land1
}
\description{
A sample of land layer at high resolution.
}
\keyword{datasets}
