% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land2}
\alias{land2}
\title{Land raster at low resolution}
\format{
\subsection{\code{land2}}{

A serialized raster.
To unpack use:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(terra)
data(land2)
land2 <- unwrap(land2)
}\if{html}{\out{</div>}}
}
}
\source{
authors
}
\usage{
land2
}
\description{
A sample of land layer at low resolution.
}
\keyword{datasets}
