% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{meanC}
\alias{meanC}
\title{Compute the mean of a numeric vector.}
\usage{
meanC(x)
}
\arguments{
\item{x}{NumericVector of values.}
}
\value{
The arithmetic mean of x.
}
\description{
Compute the mean of a numeric vector.
}
