% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwavefetch-package.R
\docType{package}
\name{rwavefetch-package}
\alias{rwavefetch}
\alias{rwavefetch-package}
\title{rwavefetch: Generate wave fetch maps}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
