% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-raster.R
\name{to_raster}
\alias{to_raster}
\title{Convert output of wave fetch functions to raster}
\usage{
to_raster(metric_matrix, extent = NULL, crs = NULL, fill_na = TRUE)
}
\arguments{
\item{metric_matrix}{any matrix returned by one of the functions of this package}

\item{extent}{an optional extent object (generated with \code{\link[terra:ext]{terra::ext()}})}

\item{crs}{an optional CRS declaration}

\item{fill_na}{if TRUE, it will fill all 0s with \code{NA}}
}
\value{
SpatRaster
}
\description{
Convert output of wave fetch functions to raster
}
\details{
Convert results to a SpatRaster.
}
\examples{
\dontrun{
to_raster(wave_fetch)
}

}
